+++ 3_Coding/TransitSys/client/src/main/java/edu/nju/umr/vo/ConstantVO.java
@@ -1,8 +1,10 @@
 package edu.nju.umr.vo;
 
+import java.math.BigDecimal;
+
 public class ConstantVO {
-	private double maxLoadPlane,maxLoadTrain,maxLoadVan,pricePlane,priceTrain,priceVan,lvEco,lvStd,lvVip;
-	public ConstantVO(double maxLoadPlane,double maxLoadTrain,double maxLoadVan,double pricePlane,double priceTrain,double priceVan,double lvEco,double lvStd,double lvVIP)
+	private BigDecimal maxLoadPlane,maxLoadTrain,maxLoadVan,pricePlane,priceTrain,priceVan,lvEco,lvStd,lvVip;
+	public ConstantVO(BigDecimal maxLoadPlane,BigDecimal maxLoadTrain,BigDecimal maxLoadVan,BigDecimal pricePlane,BigDecimal priceTrain,BigDecimal priceVan,BigDecimal lvEco,BigDecimal lvStd,BigDecimal lvVIP)
 	{
 		this.maxLoadPlane=maxLoadPlane;
 		this.maxLoadTrain=maxLoadTrain;
@@ -14,31 +16,31 @@ public class ConstantVO {
 		this.lvStd=lvStd;
 		this.lvVip=lvVIP;
 	}
-	public double getMaxLoadPlane(){
+	public BigDecimal getMaxLoadPlane(){
 		return maxLoadPlane;
 	}
-	public double getMaxLoadTrain(){
+	public BigDecimal getMaxLoadTrain(){
 		return maxLoadTrain;
 	}
-	public double getMaxLoadVan(){
+	public BigDecimal getMaxLoadVan(){
 		return maxLoadVan;
 	}
-	public double getPricePlane(){
+	public BigDecimal getPricePlane(){
 		return pricePlane;
 	}
-	public double getPriceTrain(){
+	public BigDecimal getPriceTrain(){
 		return priceTrain;
 	}
-	public double getPriceVan(){
+	public BigDecimal getPriceVan(){
 		return priceVan;
 	}
-	public double getLvEco(){
+	public BigDecimal getLvEco(){
 		return lvEco;
 	}
-	public double getLvStd(){
+	public BigDecimal getLvStd(){
 		return lvStd;
 	}
-	public double getLvVip(){
+	public BigDecimal getLvVip(){
 		return lvVip;
 	}
